backend/src/main/java/com/gabrielluciano/blog/services/TagServiceImpl.java
package com.gabrielluciano.blog.services;

import com.gabrielluciano.blog.dto.tag.TagCreateRequest;
import com.gabrielluciano.blog.dto.tag.TagUpdateRequest;
import com.gabrielluciano.blog.models.Tag;
import com.gabrielluciano.blog.repositories.TagRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class TagServiceImpl implements TagService {

    private final TagRepository tagRepository;

    @Override
    public Page<Tag> list(Pageable pageable) {
        return tagRepository.findAll(pageable);
    }

    @Override
    public Tag findById(long id) {
        return null;
    }

    @Override
    public Tag save(TagCreateRequest tagCreateRequest) {
        return null;
    }

    @Override
    public void update(TagUpdateRequest tagUpdateRequest) {

    }

    @Override
    public void deleteById(long id) {

    }
}
